requirements.txt
PyGitHub
networkx